食品詳細画面をHomeから分離
